+++ src/main/java/edu/ntnu/idatt2106/boco/controllers/UserController.java
@@ -87,6 +87,11 @@ public class UserController
     public ResponseEntity<String> deleteUser(@PathVariable long userId){
         try
         {
+            if (!tokenComponent.haveAccessTo(userId))
+            {
+                return new ResponseEntity(HttpStatus.FORBIDDEN);
+            }
+
             boolean success = userService.deleteUser(userId);
             if(!success)
             {
@@ -105,6 +110,11 @@ public class UserController
     public ResponseEntity<UserResponse> updateUser(@PathVariable("userId") long userId, @RequestBody UpdateUserRequest request) {
         try
         {
+            if (!tokenComponent.haveAccessTo(userId))
+            {
+                return new ResponseEntity(HttpStatus.FORBIDDEN);
+            }
+
             UserResponse user = userService.updateUser(userId, request);
             if (user == null)
             {
